layer.conf: Add bluez5 to DISTRO_FEATURES to bypass multiple bluez-hcidump providers error.
We have bluetooth-rfkill-event in our packagegroup and in the new
edison bsp repository, this package depends on bluez5.

bluez5 and bluez-hcidump are both providing bluez-hcidump. In order
to bypass this multiple providers error, we add bluez5 to DISTRO_FEATURES so
that the build system will set the preferred provider for this as being bluez5.
(see poky/meta/conf/distro/include/default-providers.inc for details on this)

Signed-off-by: Florin Sarbu <florin@resin.io>