Describe binary dependencies
This PR describes the binary dependencies needed to initialise certain
device types. Some devices, like the Edison, require some utilities to
be available in the PATH in order to be initialised correctly.

The CLI will attempt to read this new `initialization.dependencies`
property, will check that the tools are avaialble, and if not, print the
corresponding instructions.

See https://resinio.atlassian.net/browse/RES-2021