linux-yocto-edison.bbappend: Add aufs kernel code from git://git.code.sf.net/p/aufs/aufs3-standalone , branch aufs3.10.x
We need to add the aufs code in this way (as a big patch) and not rely
on our default way of adding the aufs patches which is implicitly done
by inheriting the kernel-resin.bbclass because the kernel we are using
contains some backported patches which have impact on adding aufs kernel
support as well as this particular aufs branch we are using needs some
changes for our kernel, changes that have not made it into the aufs tree
because aufs3 is discontinued.

So we must first revert a backported commit, we do this with the
0001-Revert-get-rid-of-s_files-and-files_lock.patch patch file, then
we add the aufs patch all in one batch by applying the aufs-builtin.patch
patch file and finally we patch the aufs code with the adapt_aufs_changes.patch
patch file.

Signed-off-by: Florin Sarbu <florin@resin.io>