resin-image.bbappend: Condition do_bootimg task to networkmanager deploy task
We revert a previous change and now we copy the kernel with initramfs and Network Manager
sample config files in the populate() function from build_hddimg task. Also, for this operation
to succeed, we need to make sure at the time we do this copying the sample config files have
already been deployed so we add a dependency to do_bootimg on Network Manager's do_deploy task.

Signed-off-by: Florin Sarbu <florin@resin.io>