meta-rust: Update to morty
We need a newer version of meta-rust in order to have a couple of
features that are missing in jethro, for example fetching of
cargo crate dependencies. The version of Rust in krogoth is also
too old to build healthdog which has dependencies which require
Rust 1.13. The easiest way to achieve this seems to be to use the
morty branch which contains all these fixes and with sufficient
backports seems to work with jethro.

Signed-off-by: Will Newton <willn@resin.io>